requirements.txt
customtkinter>=5.2.0
anthropic>=0.34.0
python-docx>=1.1.0
pdfplumber>=0.11.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
Pillow>=10.3.0
pyinstaller>=6.7.0
